backend/requirements.txt
# Dependências do Backend Python
# Para instalar: pip install -r requirements.txt

# Framework Web
Flask==3.0.0
flask-cors==4.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Gerenciamento de Variáveis de Ambiente
python-dotenv==1.0.0

# Utilitários
Pillow==10.1.0  # Manipulação de imagens (se necessário)

# Produção (opcional)
gunicorn==21.2.0  # Para deploy em produção